refactor: HandlerMappings 일급컬렉션화
