+++ mvc/src/main/java/nextstep/mvc/controller/handler/HandlerMappings.java
@@ -1,0 +1,31 @@
+package nextstep.mvc.controller.handler;
+
+import jakarta.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+public class HandlerMappings {
+
+    private final List<HandlerMapping> handlerMappings;
+
+    public HandlerMappings() {
+        this.handlerMappings = new ArrayList<>();
+    }
+
+    public void initialize() {
+        handlerMappings.forEach(HandlerMapping::initialize);
+    }
+
+    public void add(final HandlerMapping handlerMapping) {
+        handlerMappings.add(handlerMapping);
+    }
+
+    public Object getHandlerMapping(final HttpServletRequest request) {
+        return handlerMappings.stream()
+            .map(handlerMapping -> handlerMapping.getHandler(request))
+            .filter(Objects::nonNull)
+            .findFirst()
+            .orElseThrow();
+    }
+}
